Sync changes to server/blueprints. Adjust TaskDescription and ChatMessage onClick handler to be the vanilla version for the blueprint
