+++ mephisto/server/blueprints/parlai_chat/webapp/src/main.js
@@ -3,5 +3,5 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import './app.jsx';
-import './css/style.css';
+import "./app.jsx";
+import "./css/style.css";
